Lab 2 adjustments
The RectangularObject superclass was adapted to include the left and top pos as properties and create an address of type point from it.

The size of the planes is now controlled by a random number generator

The issue of the passenger window separation is now fixed.

Further adjustments where made to suite lab 2's specs, which aren't really that important