Lab 3 modifications
The position of the Airplanes as well as their size is now completely randomised. Furthermore,  buttons have been added to control the color of the plane, size/position and visibility.

The outline object class had 3 new methods added for different shapes.

A coupe of cosmetic changes has been also made, in the form of adding shadows to the dome, adding the ATC tower, adding gears to the pane as well as the cockpit windows and adding the terminal windows. Some other minor cosmetic changes where made